[snap-conversion-api] Base refactor to switch between snap capi implementations
