+++ packages/destination-actions/src/destinations/snap-conversions-api/snap-capi-properties.ts
@@ -1,8 +1,4 @@
-import { IntegrationError } from '@segment/actions-core'
 import { InputField } from '@segment/actions-core/destination-kit/types'
-import { createHash } from 'crypto'
-import { Settings } from '../snap-conversions-api/generated-types'
-import { Payload } from './reportConversionEvent/generated-types'
 
 export const CURRENCY_ISO_4217_CODES = new Set([
   'USD',
@@ -357,111 +353,5 @@ export const click_id: InputField = {
   type: 'string',
   default: {
     '@path': '$.integrations.Snap Conversions Api.click_id'
-  }
-}
-
-//Check to see what ids need to be passed depending on the event_conversion_type
-export const conversionType = (settings: Settings, event_conversion_type: String): Settings => {
-  if (event_conversion_type === 'MOBILE_APP') {
-    if (!settings?.snap_app_id || !settings?.app_id) {
-      throw new IntegrationError(
-        'If event conversion type is "MOBILE_APP" then Snap App ID and App ID must be defined',
-        'Misconfigured required field',
-        400
-      )
-    }
-    delete settings?.pixel_id
-  } else {
-    if (!settings?.pixel_id) {
-      throw new IntegrationError(
-        `If event conversion type is "${event_conversion_type}" then Pixel ID must be defined`,
-        'Misconfigured required field',
-        400
-      )
-    }
-    delete settings?.snap_app_id
-    delete settings?.app_id
-  }
-  return settings
-}
-
-export const hash = (value: string | undefined): string | undefined => {
-  if (value === undefined) return
-
-  const hash = createHash('sha256')
-  hash.update(value)
-  return hash.digest('hex')
-}
-
-const isHashedEmail = (email: string): boolean => new RegExp(/[0-9abcdef]{64}/gi).test(email)
-
-const transformProperty = (property: string, items: Array<Record<string, string | number | undefined>>): string =>
-  items
-    .map((i) =>
-      i[property] === undefined || i[property] === null
-        ? ''
-        : typeof i[property] === 'number'
-        ? (i[property] as number).toString()
-        : (i[property] as string).toString().replace(/;/g, '')
-    )
-    .join(';')
-
-export const formatPayload = (payload: Payload): Object => {
-  //Normalize fields based on Snapchat Data Hygiene https://marketingapi.snapchat.com/docs/conversion.html#auth-requirements
-  if (payload.email) {
-    //Removes all leading and trailing whitespace and converts all characters to lowercase.
-    payload.email = payload.email.replace(/\s/g, '').toLowerCase()
-  }
-
-  if (payload.phone_number) {
-    //Removes all non-numberic characters and leading zeros.
-    payload.phone_number = payload.phone_number.replace(/\D|^0+/g, '')
-  }
-
-  if (payload.mobile_ad_id) {
-    //Converts all characters to lowercase
-    payload.mobile_ad_id = payload.mobile_ad_id.toLowerCase()
-  }
-
-  let item_ids: string | undefined = undefined
-  let item_category: string | undefined = undefined
-  let brands: string[] | undefined = undefined
-
-  // if customer populates products array, use it instead of individual fields
-  const p = payload?.products
-  if (p && Array.isArray(p) && p.length > 0) {
-    item_ids = transformProperty('item_id', p)
-    item_category = transformProperty('item_category', p)
-    brands = p.map((product) => product.brand ?? '')
-  }
-
-  return {
-    event_type: payload?.event_type,
-    event_conversion_type: payload?.event_conversion_type,
-    event_tag: payload?.event_tag,
-    timestamp: Date.parse(payload?.timestamp),
-    hashed_email: isHashedEmail(String(payload?.email)) ? payload?.email : hash(payload?.email),
-    hashed_mobile_ad_id: hash(payload?.mobile_ad_id),
-    uuid_c1: payload?.uuid_c1,
-    hashed_idfv: hash(payload?.idfv),
-    hashed_phone_number: hash(payload?.phone_number),
-    user_agent: payload?.user_agent,
-    hashed_ip_address: hash(payload?.ip_address),
-    item_category: item_category ?? payload?.item_category,
-    brands: brands ?? payload?.brands,
-    item_ids: item_ids ?? payload?.item_ids,
-    description: payload?.description,
-    number_items: payload?.number_items,
-    price: payload?.price,
-    currency: payload?.currency,
-    transaction_id: payload?.transaction_id,
-    level: payload?.level,
-    client_dedup_id: payload?.client_dedup_id,
-    search_string: payload?.search_string,
-    page_url: payload?.page_url,
-    sign_up_method: payload?.sign_up_method,
-    device_model: payload?.device_model,
-    os_version: payload?.os_version,
-    click_id: payload?.click_id
   }
 }
